python_experiment/requirements.txt
imutils==0.5.4
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
